refactor: remove mocks
