+++ client/src/projects/List.tsx
@@ -1,11 +1,14 @@
-import react from 'react';
+import react, { useState } from 'react';
 import { Link } from 'react-router-dom';
 
-import { PROJECTS } from '../mocks/projects';
-
 export default function() {
+    const [projects, setProjects] = useState<any[]>([]);
+    
+    fetch('http://api.admin.localhost/project')
+        .then(res => res.json())
+        .then(res => setProjects([...res]));
 
-    const projectList = PROJECTS.map((project) => (
+    const projectList = projects.map((project) => (
         <div>{project.name}
             <span>
                 <button><Link to={`/admin/projects/${project.name}`}>edit</Link></button>
